Rename project to align with Microsoft NuGet package
