+++ Source/EventFlow.DependencyInjection/Properties/AssemblyInfo.cs
